requirements.txt
peewee
python-telegram-bot
jinja2
textual
rich-pixels

